Copy content from docs/source/commands.md to GUIDE.md
This is very helpful, because you only need to update the sphinx documentation and then it will also update GUIDE.md (which will be same). I might remove GUIDE.md at some point, but I have included it now, because it's nice for users that are checking the repository in GitHub.