Add pre-commit workflow
I wanted to include pre-commit workflow (although it is used in githooks), because just of the possibility that someone doesn't have pre-commit and also certain changes must be made directly in the GitHub website. Instead of having to check whether it's correct (pre-commit wise) I wanted to include the check in a workflow.