Remove quotes
While using quotes is generally recommended in this project they're not needed. None of the arguments take spaces or special characters. So there's no requirement and adding it here would mean that I would have to add quotes on every single variable (which I will not do). This keeps consistency.