Encapsulate methods for improved file-level structure organization
Honestly this might seem somewhat of a overkill usage of classes. But there is a reason why I'm using so many classes. Firstly I had an issue with  specifically with UtilityFunction with the suggestion being (maybe move it to another class?). So I moved the functions that had this issue. Well then I had a class in certain files, but not in others so I decided to keep it consistent and to add more classes. And using classes even it being in just one class it tends to improve the organization of the code. Improves file-level structure organization. I also made small changes to other files. I will figure out a way to create a different function than ImageValidator classes methods, if tiilux argument is not provided.