Add color and font customization support for logos tool
This commit enables customization for the logos tool written in Python. 
The file .custom_settings.yaml will have custom values. Users need to 
create this file and follow the instructions (which will be added 
later) to be able to customize logos tool color and font. Default values 
are set if file is not found or yaml has an error.