Add ColorSettings for color customization support with default colors
This commit adds color customization support for all of the other parts 
that have color. Sets DEFAULT_COLORS and checks the 
.custom_settings.yaml file for the values. If only one value is in the 
YAML file, it uses default values for the rest. Optional file, but it
cannot be kept empty. You need to add at least one. This commit also 
changes some of the other files as well.